RAG-Tech-Doc-Assistant/requirements.txt
langchain
langchain-google-genai
streamlit
faiss-cpu
pypdf
beautifulsoup4
markdown
langchain_community
python-dotenv